resolved the Incorrect use of widget
